preparing project structure for unity assets store
