--- .consulo/codeStyleSettings.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project>
-  <component name="ProjectCodeStyleSettingsManager">
-    <option name="PER_PROJECT_SETTINGS">
-      <value version="173" />
-    </option>
-    <option name="PREFERRED_PROJECT_CODE_STYLE" value="MustBe" />
-  </component>
-</project>
