--- .consulo/compiler.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="JavaCompilerConfiguration" compiler="JavacCompiler">
-    <annotation-processing>
-      <profile default="true" name="Default" enabled="false">
-        <processorPath useClasspath="true" />
-      </profile>
-    </annotation-processing>
-  </component>
-  <component name="ResourceCompilerConfiguration">
-    <resourceExtensions />
-  </component>
-</project>
