--- .consulo/copyright/profiles_settings.xml
@@ -1,5 +1,0 @@
-<component name="CopyrightManager">
-  <settings default="apache2">
-    <module2copyright />
-  </settings>
-</component>
